# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeVault API",
        "description": "OpenAPI specification for the AnimeVault backend",
        "contact": {
            "email": "support@animevault.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/me/favorites": {
            "get": {
                "tags": [
                    "User Favorites"
                ],
                "summary": "Lists the authenticated user favorite anime",
                "operationId": "apiMeFavoritesIndex",
                "responses": {
                    "200": {
                        "description": "Favorite anime list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/favorites/{anime}": {
            "put": {
                "tags": [
                    "User Favorites"
                ],
                "summary": "Adds an anime to the authenticated user favorites",
                "operationId": "apiMeFavoritesUpsert",
                "parameters": [
                    {
                        "name": "anime",
                        "in": "path",
                        "description": "Anime identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite already existed"
                    },
                    "201": {
                        "description": "Favorite created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Anime not found"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Favorites"
                ],
                "summary": "Removes an anime from the authenticated user favorites",
                "operationId": "apiMeFavoritesDelete",
                "parameters": [
                    {
                        "name": "anime",
                        "in": "path",
                        "description": "Anime identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Favorite removed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Anime not found"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/library": {
            "get": {
                "tags": [
                    "User Library"
                ],
                "summary": "Lists the authenticated user anime library",
                "operationId": "apiMeLibraryIndex",
                "responses": {
                    "200": {
                        "description": "User library list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/anime/{anime}": {
            "get": {
                "tags": [
                    "User Library"
                ],
                "summary": "Gets the authenticated user state for a specific anime",
                "operationId": "apiMeAnimeState",
                "parameters": [
                    {
                        "name": "anime",
                        "in": "path",
                        "description": "Anime identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User anime state"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Anime not found"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me/library/{anime}": {
            "put": {
                "tags": [
                    "User Library"
                ],
                "summary": "Creates or updates a library entry for the authenticated user",
                "operationId": "apiMeLibraryUpsert",
                "parameters": [
                    {
                        "name": "anime",
                        "in": "path",
                        "description": "Anime identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "watching"
                                    },
                                    "progress_episodes": {
                                        "type": "integer",
                                        "example": 18,
                                        "nullable": true
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 9,
                                        "nullable": true
                                    },
                                    "started_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "completed_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Library entry updated"
                    },
                    "201": {
                        "description": "Library entry created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Anime not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Library"
                ],
                "summary": "Deletes a library entry for the authenticated user",
                "operationId": "apiMeLibraryDelete",
                "parameters": [
                    {
                        "name": "anime",
                        "in": "path",
                        "description": "Anime identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Library entry deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Anime not found"
                    }
                },
                "security": [
                    {
                        "sanctumBearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumBearerAuth": {
                "type": "http",
                "description": "Bearer token returned by the Sanctum login or register endpoints",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}